в процессе очень сложный

+++ src/Employee.java
@@ -30,7 +30,7 @@ public class Employee {
         this.department = department;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(float salary) {
         this.salary = salary;
     }
